Add install time check for old manual
This is a first attempt at warning the user about old manual files being
present. Handling the pybind11 directory first appears to be important
for the configure step. A side effect is that must also be the home for
commands that are intended to run at the beginning of the install step.

Signed-off-by: Connor Behan <connor.behan@gmail.com>