Fixes for expand() (#179)
* Make expand() respect index position

For expand() to work on an index bracket, the factors need to have the
ImplicitIndex property, sometimes inherited through Matrix. However,
ImplicitIndex also allows a concrete convention of upper and lower
indices to be specified. When this is present for all factors, we should
use this to determine where new indices should go instead of hard coding
subscripts.

Signed-off-by: Connor Behan <connor.behan@gmail.com>

* Add expand() test

A quick verification that we get a lower index contracted with an upper
index now.

Signed-off-by: Connor Behan <connor.behan@gmail.com>

* Fix egregious typos

Some working lines were overwritten with non-working lines when I
changed the tab settings.

Signed-off-by: Connor Behan <connor.behan@gmail.com>