Prefer last indices
Is (B A)_{c a} a valid answer for (B)_{b c} (A)_{a b}? It isn't if "a"
and "c" are indices on the same footing as "b". However, they could be
pure labels which would turn the above into a commutative dot product.
To avoid this, we should assume that indices go outside the bracket
while labels go inside. It is therefore not enough to begin the
contraction on B's last participating index. We should begin with a last
index that is also participating. Sometimes this does not exist so we
keep the old logic in place. Garbage in / garbage out is better than a
crash.

Signed-off-by: Connor Behan <connor.behan@gmail.com>