Recognize different ways of being traceless
If the Trace property is defined with respect to an index set, Traceless
should be as well. An advantage of this is that canonicalise() no longer
has to demand an explicit form. This patch makes it get rid of vanishing
traces even when they act on implicit indices. Of course, this is just a
start. Tensor product representations are common enough that all classes
derived from TableauBase should eventually be modified in this way.

Signed-off-by: Connor Behan <connor.behan@gmail.com>