Use cyclicity of the trace
Analyzing an expression is easiest if all n! orderings of a given
product produce the same output when passed to sort_product().
Unfortunately, this is very hard to achieve in a CAS for non-commuting
objects. Consider matrices A and C with a scalar B. CAB and BCA are
equal but the first is "sorted" because B comes after A and the second
is "sorted" because B comes before C.

An opportunity for making this better occurs when the expression is
inside a trace. Then, B can be pulled out by untrace() and the rest can
be sorted with cyclicity.

Signed-off-by: Connor Behan <connor.behan@gmail.com>