Allow combine() to accept a trace
If we add an optional argument, we can allow the user's operators with a
Trace property to be used whenever traces are generated by combine().
This is not strictly necessary as everyone knows what (A)_{a a} means
but it can add some flexibility in notation and also serve as a warm-up
for transpose.

Also, we should remember that the most sensible use of an indexbracket
is to enclose objects that have the ImplicitIndex property.

Signed-off-by: Connor Behan <connor.behan@gmail.com>