Fix tests
To fix tests that were broken by the most recent commits, we should
remember to check that an Indices property was indeed found. A common
idiom in Cadabra is to not declare it when doing simple things. Also, we
need to account for products of traceless matrices that have no parent.
Even when there is a parent, we need to let its multiplier be propagated
from the child.

Signed-off-by: Connor Behan <connor.behan@gmail.com>