Perform a full check with cyclicity
The factor that comes alphabetically first in a trace might appear
multiple times. Therefore, cycling until this factor comes first does
not always lead to the same ordering. This moves to a better algorithm
by looking at all orderings and breaking ties with the next factor. My
old code was also bad because it called should_swap() on non-adjacent
iterators.

Signed-off-by: Connor Behan <connor.behan@gmail.com>