Recognize different ways of being antisymmetric
This makes the EpsilonTensor property perform a few sanity checks and
remember the name of the index set. This allows us to avoid joining
epsilons of different types. Joining gammas seems less error-prone
because of the implicit matrix multiplication. However, it cannot hurt
to add the same logic to GammaMatrix which is the other property derived
from AntiSymmetric.

Signed-off-by: Connor Behan <connor.behan@gmail.com>