Use index set name less for AntiSymmetric
The current function of AntiSymmetric is to quickly make a pattern
anti-symmetric in all of its indices no matter what sets they come from.
Making this property more specific would be a challenge. Consider a
tensor which is symmetric x anti-symmetric x hook-diagram. If we allow
this to be assigned by stacking calls to Symmetric, AntiSymmetric and
TableauSymmetry, all three properties will have to be queried by all
three get_tab() methods. This will either require a lot of error
trapping or a clear convention for who wins in a conflict. For now, we
should remove index_set_name from AntiSymmetric.

EpsilonTensor and GammaMatrix are exceptions which need to know the set
name because of their join algorithms. Fortunately, we can just do this
at run time.

Signed-off-by: Connor Behan <connor.behan@gmail.com>