Use index set name more for Traceless
We need to allow tensors to be traceless with respect to many index sets
at once. This patch does so but only when the set names are specified in
a single call to Traceless. To improve on things further, we should
figure out which properties benefit from being specified in many steps
and change the property injector to not overwrite in these cases.
Perhaps it should still overwrite when it finds an empty argument list.

Signed-off-by: Connor Behan <connor.behan@gmail.com>