Collect values in a different way
One of the other tests indicates that there actually is supposed to be a
difference between Integer and the values keyword. The latter is needed
to make the evaluate algorithm work. Therefore we should still attempt
to collect values in Indices but failures should not be fatal and
Integer (which we apply no matter what) should do the heavy lifting.

Signed-off-by: Connor Behan <connor.behan@gmail.com>