Replace Greek letters entered as unicode instead of TeX
People doing quick and dirty things might want to create a new
expression by copying and pasting from the output of a previous
algorithm. The asymmetry here is that input typically has TeX while
output typically has unicode. We can solve this by performing a
replacement in the parser. It is then natural to copy greekmap from
DisplayTerminal to Symbols (and add varphi to it).

In case there are users who input both TeX and unicode and rely on them
being distinct, this could be put behind a command-line argument or
environment variable.

Signed-off-by: Connor Behan <connor.behan@gmail.com>