Associate window with GtkApplication.
When a NotebookWindow is created, it is not mapped to the Cadabra
GtkApplication. This causes window managers to fail to track windows by
the app name, leading to issues such as a generic icon showing instead
of the cadabra app icon (for example on GNOME shell) and multiple
cadabra windows showing up on the dash bar without being grouped by
application. Fixed by this commit by calling
`GtkApplication::add_window` in the NotebookWindow constructor.